+++ server/node/bootstrap.cjs
@@ -1,0 +1,63 @@
+const http = require('http');
+const path = require('path');
+const fs = require('fs/promises');
+const https = require('https');
+const { sslPath } = require('./config.cjs');
+
+async function getHttpsOptions() {
+
+    const keyPath = path.join(sslPath, 'server.key');
+    const certPath = path.join(sslPath, 'server.crt');
+
+    try {
+
+        await fs.access(keyPath);
+        await fs.access(certPath);
+
+        const [key, cert] = await Promise.all([
+            fs.readFile(keyPath),
+            fs.readFile(certPath)
+        ]);
+
+        return { key, cert };
+
+    } catch (error) {
+        console.error('[Server] SSL setup errors:', error.message);
+        console.log('[Server] Start the server with HTTP instead of HTTPS...');
+        return null;
+    }
+}
+
+async function startServer(app, proxyStreamHandlers) {
+    try {
+
+        const port = process.env.PORT || 6001;
+        const httpsOptions = await getHttpsOptions();
+        let server = null;
+
+        if (httpsOptions) {
+            // HTTPS
+            server = https.createServer(httpsOptions, app);
+            proxyStreamHandlers.setupProxyStreamWebSocket(server);
+            server.listen(port, () => {
+                console.log("[Server] HTTPS server is running.");
+                console.log(`[Server] https://localhost:${port}/`);
+            });
+        } else {
+            // HTTP
+            server = http.createServer(app);
+            proxyStreamHandlers.setupProxyStreamWebSocket(server);
+            server.listen(port, () => {
+                console.log("[Server] HTTP server is running.");
+                console.log(`[Server] http://localhost:${port}/`);
+            });
+        }
+    } catch (error) {
+        console.error('[Server] Failed to start server :', error);
+        process.exit(1);
+    }
+}
+
+module.exports = {
+    startServer,
+};
