+++ server/node/server.cjs
@@ -3,13 +3,12 @@ const app = express();
 if (process.env.TRUST_PROXY) {
     app.set('trust proxy', Number(process.env.TRUST_PROXY) || process.env.TRUST_PROXY);
 }
-const http = require('http');
 const path = require('path');
 const htmlparser = require('node-html-parser');
-const { writeFileSync } = require('fs');
 const fs = require('fs/promises')
 const { createAuthHelpers, createLimiters } = require('./auth.cjs');
-const { authCodePath, sslPath } = require('./config.cjs');
+const { startServer } = require('./bootstrap.cjs');
+const { registerOAuthRoutes } = require('./routes-oauth.cjs');
 const { registerProxyRoutes } = require('./routes-proxy.cjs');
 const { createProxyStreamHandlers } = require('./routes-proxy-stream.cjs');
 const { registerStorageRoutes } = require('./routes-storage.cjs');
@@ -18,9 +17,6 @@ app.use(express.static(path.join(process.cwd(), 'dist'), {index: false}));
 app.use(express.json({ limit: '100mb' }));
 app.use(express.raw({ type: 'application/octet-stream', limit: '100mb' }));
 app.use(express.text({ limit: '100mb' }));
-const {pipeline} = require('stream/promises')
-const https = require('https');
-const openid = require('openid-client');
 
 const state = createServerState();
 const {
@@ -46,74 +42,6 @@ const proxyStreamHandlers = createProxyStreamHandlers({
     state,
     authHelpers,
 });
-
-async function forwardUpstreamResponse(originalResponse, res) {
-    const head = new Headers(originalResponse.headers);
-    head.delete('content-security-policy');
-    head.delete('content-security-policy-report-only');
-    head.delete('clear-site-data');
-    head.delete('Cache-Control');
-    head.delete('Content-Encoding');
-
-    const contentType = (head.get('content-type') || '').toLowerCase();
-    const isSSE = contentType.includes('text/event-stream');
-    if (isSSE) {
-        head.set('Cache-Control', 'no-cache, no-transform');
-        head.set('Connection', 'keep-alive');
-        head.set('X-Accel-Buffering', 'no');
-        head.delete('content-length');
-    }
-
-    const headObj = {};
-    for (const [k, v] of head) {
-        headObj[k] = v;
-    }
-
-    res.header(headObj);
-    res.status(originalResponse.status);
-
-    if (!originalResponse.body) {
-        res.end();
-        return;
-    }
-
-    if (!isSSE) {
-        await pipeline(originalResponse.body, res);
-        return;
-    }
-
-    const reader = originalResponse.body.getReader();
-
-    const onClose = () => {
-        reader.cancel().catch(() => {});
-    };
-    res.on('close', onClose);
-
-    if (typeof res.flushHeaders === 'function') {
-        res.flushHeaders();
-    }
-
-    try {
-        while (!res.writableEnded) {
-            const { done, value } = await reader.read();
-            if (done) {
-                break;
-            }
-            if (value && value.length > 0) {
-                res.write(Buffer.from(value));
-            }
-        }
-    } catch (error) {
-        if (!res.writableEnded) {
-            throw error;
-        }
-    } finally {
-        res.off('close', onClose);
-        if (!res.writableEnded) {
-            res.end();
-        }
-    }
-}
 
 app.get('/', async (req, res, next) => {
 
@@ -149,161 +77,9 @@ registerStorageRoutes(app, {
     authHelpers,
 });
 
-app.get('/api/oauth_login', async (req, res) => {
-    const redirect_uri = (new URL (req.url)).host + '/api/oauth_callback'
-
-    if(!redirect_uri){
-        res.status(400).send({ error: 'redirect_uri is required' });
-        return
-    }
-    if(!state.oauthData.client_id || !state.oauthData.client_secret){
-        const discovery = await openid.discovery('https://account.sionyw.com/','','');
-        state.oauthData.config = discovery;
-
-        //oauth dynamic client registration
-        //https://datatracker.ietf.org/doc/html/rfc7591
-
-        const serverMeta = discovery.serverMetadata()
-        //since we can't find a good library to do this, we will do it manually
-        const registrationResponse = await fetch(serverMeta.registration_endpoint, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json',
-                'Authorization': 'Bearer ' + (serverMeta.registration_access_token || '')
-            },
-            body: JSON.stringify({
-                client_id: state.oauthData.client_id,
-                client_secret: state.oauthData.client_secret,
-                redirect_uris: [redirect_uri],
-                response_types: ['code'],
-                grant_types: ['authorization_code'],
-                scope: 'risuai',
-                token_endpoint_auth_method: 'client_secret_basic',
-                client_name: 'Risuai Node Server',
-            })
-        });
-
-        if(registrationResponse.status === 201 || registrationResponse.status === 200){
-            const registrationData = await registrationResponse.json();
-            state.oauthData.client_id = registrationData.client_id;
-            state.oauthData.client_secret = registrationData.client_secret;
-            discovery.clientMetadata().client_id = state.oauthData.client_id;
-            discovery.clientMetadata().client_secret = state.oauthData.client_secret;
-        }
-        else{
-            console.error('[Server] OAuth2 dynamic client registration failed:', registrationResponse.statusText);
-            res.status(500).send({ error: 'OAuth2 client registration failed' });
-            return
-        }
-
-
-        //now lets request
-
-        let code_verifier = openid.randomPKCECodeVerifier();
-        let code_challenge = await openid.calculatePKCECodeChallenge(code_verifier);
-
-        state.oauthData.code_verifier = code_verifier;
-        let redirectTo = openid.buildAuthorizationUrl(state.oauthData.config, {
-            redirect_uri,
-            code_challenge,
-            code_challenge_method: 'S256',
-            scope: 'risuai',
-        })
-
-        res.redirect(redirectTo.toString());
-
-        return;
-
-    }
-    
-    res.status(500).send({ error: 'OAuth2 login failed' });
-});
-
-app.get('/api/oauth_callback', async (req, res) => {
-
-    //since this is a callback we don't need to check password
-
-    const params = (new URL(req.url, `http://${req.headers.host}`)).searchParams;
-    const code = params.get('code');
-
-    if(!code){
-        res.status(400).send({ error: 'code is required' });
-        return
-    }
-    if(!state.oauthData.client_id || !state.oauthData.client_secret || !state.oauthData.code_verifier){
-        res.status(400).send({ error: 'OAuth2 not initialized' });
-        return
-    }
-
-    let tokens = await openid.authorizationCodeGrant(
-        state.oauthData.config,
-        getCurrentUrl(),
-        {
-            pkceCodeVerifier: state.oauthData.code_verifier,
-        },
-    )
-
-    writeFileSync(authCodePath, tokens.access_token, 'utf-8')
-
-    res.send(tokens)
-            
-})
-
-async function getHttpsOptions() {
-
-    const keyPath = path.join(sslPath, 'server.key');
-    const certPath = path.join(sslPath, 'server.crt');
-
-    try {
- 
-        await fs.access(keyPath);
-        await fs.access(certPath);
-
-        const [key, cert] = await Promise.all([
-            fs.readFile(keyPath),
-            fs.readFile(certPath)
-        ]);
-       
-        return { key, cert };
-
-    } catch (error) {
-        console.error('[Server] SSL setup errors:', error.message);
-        console.log('[Server] Start the server with HTTP instead of HTTPS...');
-        return null;
-    }
-}
-
-async function startServer() {
-    try {
-      
-        const port = process.env.PORT || 6001;
-        const httpsOptions = await getHttpsOptions();
-        let server = null;
-
-        if (httpsOptions) {
-            // HTTPS
-            server = https.createServer(httpsOptions, app);
-            proxyStreamHandlers.setupProxyStreamWebSocket(server);
-            server.listen(port, () => {
-                console.log("[Server] HTTPS server is running.");
-                console.log(`[Server] https://localhost:${port}/`);
-            });
-        } else {
-            // HTTP
-            server = http.createServer(app);
-            proxyStreamHandlers.setupProxyStreamWebSocket(server);
-            server.listen(port, () => {
-                console.log("[Server] HTTP server is running.");
-                console.log(`[Server] http://localhost:${port}/`);
-            });
-        }
-    } catch (error) {
-        console.error('[Server] Failed to start server :', error);
-        process.exit(1);
-    }
-}
+registerOAuthRoutes(app, state);
 
 (async () => {
     proxyStreamHandlers.startProxyStreamGc();
-    await startServer();
+    await startServer(app, proxyStreamHandlers);
 })();
